Optimize Rust
By using an optimized iterator, Rust can be faster than Zig.

Not sure if you want to allow 3rd party dependencies, but num_iter is written in pure Rust.

This also cleans up the writing in Rust and changes the end time to be calculated within the callback for Zig and Rust, making both faster.